refactor: Move navigation parameters logic to home screen
